+++ CHANGELOG.md
@@ -27,6 +27,7 @@ All notable changes to this project will be documented in this file.
 - Speed Highway's shutter and elevator will open for Amy and Big when the player gets the Employee Card.
 - Big's Speed Highway entrance is at the city hall, he still needs the Employee Card to enter.
 - Knuckles' barricade won't spawn if the level for that entrance is not a level for him.
+- Big, Knuckles and Tails can use Gamma's Final Egg entrance.
 
 ### Changed